+++ app/templates/static/js/events.js
@@ -1,3 +1,18 @@
+function notifyMe(caller) {
+	if (!("Notification" in window)) {
+		//alert("This browser does not support desktop notification");
+	} else if (Notification.permission === "granted") {
+		var notification = new Notification("Studio-Link call from: " + caller);
+		notification.onclick = function() {
+			window.focus();
+		};
+	} else if (Notification.permission !== 'denied') {
+		Notification.requestPermission(function (permission) {
+			// Nothing todo
+		});
+	}
+}
+
 //Long Polling Call Events
 function checkCallEvents()
 {
@@ -10,6 +25,7 @@ function checkCallEvents()
     }).done(function(result) {
 
         if (result.INCOMING) {
+	    notifyMe(result.INCOMING);
             bootbox.confirm("Incoming call from '" + result.INCOMING  + "', accept?", 
                 function(result) {
                     if (result) {